requirements.txt
# Core
typecast-python>=0.1.3

# YAML parsing
PyYAML>=6.0

# Optional (recommended)
python-dotenv>=1.0.0
